[Common] Getting rid of warning 4996.

+++ Server_Common/src/db_expiration_engine.cpp
@@ -28,7 +28,7 @@ Engine::~Engine()
 void Engine::ScheduleEverySeconds(time_t seconds_interval)
 {  
   expiration_task = new Task(database_handler_);
-  schedule(expiration_task, 0, seconds_interval * 1000);
+  schedule(expiration_task, 0, static_cast<long>(seconds_interval * 1000));
 
   logger_.information("[expiration/Engine] Scheduled for every " +
     NumberFormatter::format(seconds_interval) + " minutes.");
